+++ src/main.rs
@@ -912,7 +912,8 @@ async fn main_internal() -> Result <(), String> {
                 if !is_json {
                     format!("Error: {}", e)
                 } else {
-                    let err: serde_json::Value = json!(e);
+                    let err: serde_json::Value = serde_json::from_str(&e)
+                        .unwrap_or(serde_json::Value::String(e));
                     let res = json!({"Error": err});
                     serde_json::to_string_pretty(&res)
                         .unwrap_or("{{ \"JSON serialization error\" }}".to_string())